refactor(shifu): extract shifu-core, the role framework library
Split the launcher into a thin binary plus shifu-core, an unpublished
workspace library carrying the parts of the shifu role every bearer
needs: pinned-tool bootstrap (tools.rs lineage), host probing helpers,
and terminal styling. The launcher keeps process ownership (dispatch,
delegation, exit policy) and adapts the library's Result-based
bootstrap to its die-with-127 contract.

Zero behavior change: --help / --version / self-version / doctor
outputs and exit codes are byte-identical before and after the split.
The release pin surface stays one crate wide: shifu-core is never
published or tagged, its version is not lerna-synced, and
sync-shifu-version.mjs / release-shifu.yml are untouched.
