+++ crates/Cargo.toml
@@ -5,11 +5,13 @@
 # Every in-repo Rust component lives here as a workspace member; see
 # docs/rust-adoption.md for the adoption paradigm (source layout, consumption
 # modes, distribution, and the selective-exercise discipline). The first member
-# is shifu, the self-contained dev/build launcher.
+# is shifu, the self-contained dev/build launcher; shifu-core is the
+# unpublished library behind it (the shifu role framework — bootstrap + probe —
+# shared with the product trunk per ADR-0046).
 
 [workspace]
 resolver = "2"
-members = ["shifu"]
+members = ["shifu", "shifu-core"]
 # host-spike is a throwaway feasibility probe (rust host shell → libkungfu →
 # embedded libpython/libnode). It builds against a sibling worktree's core
 # artifacts, so it must not enter the workspace CI gate or the release matrix;
